Fix sign-compare compiler warnings in d directory
Lots of ffi.d changes since many of the functions return an
ffi_status, which is an enum.

+++ M2/Macaulay2/d/scclib.c
@@ -411,13 +411,13 @@ int system_fileMode(M2_string name) {
   struct stat buf;
   int r = stat(cname,&buf);
   freemem(cname);
-  return r == ERROR ? -1 : buf.st_mode & ~S_IFMT;
+  return r == ERROR ? -1 : (int)(buf.st_mode & ~S_IFMT);
 }
 
 int system_fileModeFD(int fd) {
   struct stat buf;
   int r = fstat(fd,&buf);
-  return r == ERROR ? -1 : buf.st_mode & ~S_IFMT;
+  return r == ERROR ? -1 : (int)(buf.st_mode & ~S_IFMT);
 }
 
 int system_chmod(M2_string name,int mode) {
